♻️ refactoring to allAccounts api method considering shared code with account api
